LayoutTests/http/tests/security/contentSecurityPolicy/connect-src-websocket-blocked-expected.txt

CONSOLE MESSAGE: line 1: Refused to load connect from 'ws://localhost:8880/websocket/tests/hybi/echo' because of Content-Security-Policy.

Pass
